+++ scopes/react/component-highlighter/label/label.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import classNames from 'classnames';
-import Tippy, { TippyProps } from '@tippyjs/react/headless';
 import { ComponentMetaHolder } from '@teambit/react.ui.highlighter.component-metadata.bit-component-meta';
 
 import styles from './label.module.scss';
 import { ComponentStrip } from './component-strip';
+import { OtherComponentsPopper } from './other-components';
 
 export interface LabelProps extends React.HTMLAttributes<HTMLDivElement> {
   components: ComponentMetaHolder[];
@@ -33,50 +33,5 @@ export function Label({ components, ...props }: LabelProps) {
         )}
       </ComponentStrip>
     </OtherComponentsPopper>
-  );
-}
-
-type OtherComponentsProps = {
-  components: ComponentMetaHolder[];
-  start?: number;
-  end?: number;
-} & TippyProps;
-
-// a popper ("tooltip") that shows the additional React Components related to this dom element
-export function OtherComponentsPopper({
-  components,
-  children,
-  placement = 'bottom',
-  interactive = true,
-  start,
-  end = -1,
-  ...tippyProps
-}: OtherComponentsProps) {
-  const content = (
-    <>
-      {components
-        .slice(start, end)
-        .reverse()
-        .map((comp, idx) => (
-          <ComponentStrip key={idx} component={comp} />
-        ))}
-    </>
-  );
-
-  return (
-    <Tippy
-      placement={placement}
-      interactive={interactive}
-      {...tippyProps}
-      // second parameter "content" is always undefined, use content inline
-      // https://github.com/atomiks/tippyjs-react/issues/341
-      render={(attrs) => (
-        <div {...attrs} className={styles.othersContainer}>
-          {content}
-        </div>
-      )}
-    >
-      {children}
-    </Tippy>
   );
 }
